% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{data_preprocess}
\alias{data_preprocess}
\title{Preprocess Data}
\usage{
data_preprocess(expr, anno_processed)
}
\arguments{
\item{expr}{A matrix of input data. Each row represents a gene and each column represents a cell.}

\item{anno_processed}{A list of processed annotation information that consists of the design matrix and the weight matrix for marker genes.}
}
\value{
A list contains processed expression matrix, design matrix, and weight matrix.
}
\description{
This function validates and preprocesses the input data for the downstream analysis.
}
