% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_matrix_builder.R
\name{design_matrix_builder}
\alias{design_matrix_builder}
\title{Design Matrix Builder}
\usage{
design_matrix_builder(anno, weight)
}
\arguments{
\item{anno}{A matrix or data frame that contains marker genes specified for cell types of interest.}

\item{weight}{The default weight assigned to marker genes.}
}
\value{
A list contains processed design matrix and weight matrix.
}
\description{
Builds the design matrix required by \code{update_func} based on user input.
}
