% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_C.R
\name{update_C}
\alias{update_C}
\title{Update Cluster}
\usage{
update_C(dat, mu_mat, designmat, dat_rt_t = NULL, rt_sq = NULL)
}
\arguments{
\item{dat}{A matrix of input data.}

\item{mu_mat}{Center estimates from \code{update_mu}}

\item{designmat}{An indicator variable matrix records specified marker genes of each cell type.}
}
\description{
Updates cluster assignments based on center estimates from \code{update_mu}
}
\keyword{internal}
