% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_func.R
\name{update_func}
\alias{update_func}
\title{Update Function}
\usage{
update_func(
  dat,
  design_mat,
  weightmat,
  unknown_threshold1 = 0,
  unknown_threshold2 = 0.05,
  max_iter = 100
)
}
\arguments{
\item{dat}{A matrix of input data.}

\item{design_mat}{An indicator variable matrix records specified marker genes of each cell type.}

\item{weightmat}{A matrix of weights assigned to each marker gene.}

\item{unknown_threshold1}{The parameter determines undecided cells cutoff. The default value is 0.}

\item{unknown_threshold2}{The parameter determines whether undecided cells are further processed. The default value is 0.05.}

\item{max_iter}{The maximum number of iterations for the algorithm to update parameters. The default value is 100.}
}
\value{
A list contains parameter estimates, type assignments, and the corresponding cost.
}
\description{
Implements the scSorter method by iteratively running \code{update_mu} and \code{update_C}.
}
