% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mu.R
\name{update_mu}
\alias{update_mu}
\title{Mu Update}
\usage{
update_mu(dat, designmat, clus)
}
\arguments{
\item{dat}{A matrix of input data.}

\item{designmat}{An indicator variable matrix records marker genes of each pre-specified cell type.}

\item{clus}{A vector of cluster assignment.}
}
\value{
A matrix of parameter estimates.
}
\description{
Solves mu and delta given sample cluster assignment.
}
