% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_temp.R
\name{xnormalize_scData}
\alias{xnormalize_scData}
\title{Normalize scRNA-seq Data}
\usage{
xnormalize_scData(expr)
}
\arguments{
\item{expr}{A matrix of input scRNA-seq data. Rows correspond to genes and columns correpond to cells.}
}
\value{
A matrix of normalized expression data.
}
\description{
Normalize scRNA-seq data. Please only use this function when you do not have access to Seurat package. More details are available in the vignette of this package.
}
