% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_calc.R
\name{cost_func}
\alias{cost_func}
\title{Cost Function}
\usage{
cost_func(dat, clus, mu, designmat)
}
\arguments{
\item{dat}{A matrix of input data.}

\item{clus}{A vector of predicted cell types.}

\item{mu}{Parameter estimates from \code{update_mu}.}

\item{designmat}{An indicator variable matrix records specified marker genes of each cell type.}
}
\description{
Calculates the cost.
}
