% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSorter.R
\name{RunScSorter}
\alias{RunScSorter}
\title{Run scSorter on a Seurat object}
\usage{
RunScSorter(
  object,
  anno,
  layer = NULL,
  assay = NULL,
  top_vf = 2000,
  min_pct = 0.1,
  set_ident = TRUE,
  name = "scSorter_cell_type",
  ...
)
}
\arguments{
\item{object}{A Seurat object containing the expression data.
`FindVariableFeatures` should be run on the Seurat object before using this function.}

\item{anno}{A matrix or data frame that contains marker genes specified for cell types of interest.
It should contain three columns named "Type", "Marker", and "Weight" that records the name and weight of marker genes specified for each cell type.
"Weight" column is optional. If it is not specified, the \code{default_weight} will be applied to all marker genes.}

\item{layer}{The name of the layer in the Seurat object to use for expression data. If NULL, the default data will be used. The default value is NULL.}

\item{assay}{The name of the assay in the Seurat object to use for expression data. If NULL, the default assay will be used. The default value is NULL.}

\item{top_vf}{The number of top variable features to use for scSorter. If NULL, all variable features will be used. The default value is NULL.}

\item{min_pct}{The minimum fraction of cells that must have non-zero expression for a gene to be retained. The default value is 0.1.}

\item{set_ident}{A logical value indicating whether to set the predicted cell types as the active identity class in the Seurat object. The default value is TRUE.}

\item{name}{The name of the metadata column to store the predicted cell types. The default value is "scSorter_cell_type".}

\item{...}{Additional arguments to pass to the scSorter function.}
}
\value{
The Seurat object with the predicted cell types added to the metadata.
}
\description{
This function runs the scSorter method on a Seurat object.
}
